fix: update form and validation data on sync
When `initialValue` is set in the annotations, the value is passed to the
controller but is not added to the `formValueMap`. With this behavior cannot
be submitted or validated properly. To avoid this situation we update
the form data when view is synced with viewmodel.

In addition, if `autoTextFieldValidation` is true, we also update the
validation data.